Fix #13793: Clip paths in a way that let's the dashes stay in place.
git-svn-id: http://josm.openstreetmap.de/svn/trunk@11144 0c6e7542-c601-0410-84e7-c038aed88b3b

+++ src/org/openstreetmap/josm/gui/MapViewState.java
@@ -552,7 +552,7 @@ public final class MapViewState implements Serializable {
          * @return The new point
          * @since 10874
          */
-        public MapViewPoint interpolate(MapViewPoint p1, int i) {
+        public MapViewPoint interpolate(MapViewPoint p1, double i) {
             return new MapViewViewPoint((1 - i) * getInViewX() + i * p1.getInViewX(), (1 - i) * getInViewY() + i * p1.getInViewY());
         }
     }
